Fixed landscape h/f
The changes fix the header and footer and rotate landscape pages in the resdoc. These changes require the information (that is to be placed on the landscape page) to be put within the landscapepage environment using \clearpage \begin{landscapepage} ....
\end{landscapepage} \clearpage.

If using csasdown::csas_table, landscape must be set to FALSE (i.e. landscape = FALSE)